requirements.txt
numpy==1.20.1
Pillow==8.1.2
Shapely==1.7.1
